lambda/requirements.txt
cassandra-driver==3.28.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3